align(google): consolidate Google ViewModels under Models/Google/
