+++ src/Humans.Web/Models/Google/GoogleSyncAuditViewModels.cs
@@ -1,0 +1,26 @@
+using Humans.Domain.Enums;
+
+namespace Humans.Web.Models.Google;
+
+public class GoogleSyncAuditEntryViewModel
+{
+    public AuditAction Action { get; set; }
+    public string Description { get; set; } = string.Empty;
+    public string? UserEmail { get; set; }
+    public string? Role { get; set; }
+    public GoogleSyncSource? SyncSource { get; set; }
+    public DateTime OccurredAt { get; set; }
+    public bool? Success { get; set; }
+    public string? ErrorMessage { get; set; }
+    public string? ResourceName { get; set; }
+    public Guid? ResourceId { get; set; }
+    public Guid? RelatedEntityId { get; set; }
+}
+
+public class GoogleSyncAuditListViewModel
+{
+    public List<GoogleSyncAuditEntryViewModel> Entries { get; set; } = [];
+    public string Title { get; set; } = string.Empty;
+    public string? BackUrl { get; set; }
+    public string? BackLabel { get; set; }
+}
