+++ src/Humans.Web/Models/Google/SyncSettingsViewModels.cs
@@ -1,6 +1,6 @@
 using Humans.Domain.Enums;
 
-namespace Humans.Web.Models;
+namespace Humans.Web.Models.Google;
 
 public class SyncSettingsViewModel
 {
