+++ src/Humans.Web/Views/Google/SyncSettings.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.SyncSettingsViewModel
+@model Humans.Web.Models.Google.SyncSettingsViewModel
 @{
     ViewData["Title"] = "Sync Settings";
 }
