+++ src/Humans.Web/Views/Shared/GoogleSyncAudit.cshtml
@@ -1,4 +1,4 @@
-@model Humans.Web.Models.GoogleSyncAuditListViewModel
+@model Humans.Web.Models.Google.GoogleSyncAuditListViewModel
 @{
     ViewData["Title"] = Model.Title;
 }
