+++ src/Humans.Web/Views/_ViewImports.cshtml
@@ -2,6 +2,7 @@
 @using Humans.Web.Extensions
 @using Humans.Web.Helpers
 @using Humans.Web.Models
+@using Humans.Web.Models.Google
 @using Humans.Web.ViewComponents
 @using Humans.Domain.Entities
 @using Humans.Domain.Enums
